requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
terminaltables3~=4.0.0